move out common component
