+++ src/applications/gi-sandbox/components/SearchResults.jsx
@@ -1,63 +1,12 @@
-import React, { useState } from 'react';
+import React from 'react';
 import SearchResultCard from './SearchResultCard';
-import SearchBenefits from './SearchBenefits';
-import AccordionDropdown from './AccordionDropdown';
-import LearnMoreLabel from './LearnMoreLabel';
-import RadioButtons from './RadioButtons';
-import { connect } from 'react-redux';
-import { updateEligibilityAndFilters, showModal } from '../actions';
+import TuitionAndHousingEstimates from './search/TuitionAndHousingEstimates';
+import LoadingIndicator from '@department-of-veterans-affairs/component-library/LoadingIndicator';
 
-export function SearchResults({
-  search,
-  eligibility,
-  dispatchUpdateEligibilityAndFilters,
-  dispatchShowModal,
-}) {
-  const [giBillChapter, setGiBillChapter] = useState(eligibility.giBillChapter);
-  const [openName, setOpenName] = useState('');
-
-  const [militaryStatus, setMilitaryStatus] = useState(
-    eligibility.militaryStatus,
-  );
-
-  const [spouseActiveDuty, setSpouseActiveDuty] = useState(
-    eligibility.spouseActiveDuty,
-  );
-
-  const [cumulativeService, setCumulativeService] = useState(
-    eligibility.cumulativeService,
-  );
-
-  const [enlistmentService, setEnlistmentService] = useState(
-    eligibility.enlistmentService,
-  );
-
-  const [eligForPostGiBill, setEligForPostGiBill] = useState(
-    eligibility.eligForPostGiBill,
-  );
-
-  const [numberOfDependents, setNumberOfDependents] = useState(
-    eligibility.numberOfDependents,
-  );
-
-  const [onlineClasses, setOnlineClasses] = useState(eligibility.onlineClasses);
-
-  const updateStore = () => {
-    dispatchUpdateEligibilityAndFilters({
-      militaryStatus,
-      spouseActiveDuty,
-      giBillChapter,
-      cumulativeService,
-      enlistmentService,
-      eligForPostGiBill,
-      numberOfDependents,
-      onlineClasses,
-    });
-  };
-
-  const handleAccordionDropdownOpen = openedName => {
-    setOpenName(openedName);
-  };
+export default function SearchResults({ search }) {
+  if (search.inProgress) {
+    return <LoadingIndicator message="Loading search results..." />;
+  }
 
   return (
     <>
@@ -68,55 +17,7 @@ export function SearchResults({
             <strong>{search.query.name}</strong>'
           </p>
           <div className="usa-width-one-third">
-            <div>
-              <AccordionDropdown
-                button="Update tuition and housing estimates"
-                buttonLabel="Update results"
-                buttonOnClick={updateStore}
-                name="benefitEstimates"
-                openName={openName}
-                onOpen={handleAccordionDropdownOpen}
-                displayCancel
-              >
-                <SearchBenefits
-                  cumulativeService={cumulativeService}
-                  dispatchShowModal={dispatchShowModal}
-                  eligForPostGiBill={eligForPostGiBill}
-                  enlistmentService={enlistmentService}
-                  giBillChapter={giBillChapter}
-                  militaryStatus={militaryStatus}
-                  numberOfDependents={numberOfDependents}
-                  spouseActiveDuty={spouseActiveDuty}
-                  setCumulativeService={setCumulativeService}
-                  setEligForPostGiBill={setEligForPostGiBill}
-                  setEnlistmentService={setEnlistmentService}
-                  setNumberOfDependents={setNumberOfDependents}
-                  setGiBillChapter={setGiBillChapter}
-                  setMilitaryStatus={setMilitaryStatus}
-                  setSpouseActiveDuty={setSpouseActiveDuty}
-                />
-                <RadioButtons
-                  label={
-                    <LearnMoreLabel
-                      text="Will you be taking any classes in person?"
-                      onClick={() =>
-                        dispatchShowModal('onlineOnlyDistanceLearning')
-                      }
-                      ariaLabel="Learn more about how we calculate your housing allowance based on where you take classes"
-                    />
-                  }
-                  name="inPersonClasses"
-                  options={[
-                    { value: 'no', label: 'Yes' },
-                    { value: 'yes', label: 'No' },
-                  ]}
-                  value={onlineClasses}
-                  onChange={e => {
-                    setOnlineClasses(e.target.value);
-                  }}
-                />
-              </AccordionDropdown>
-            </div>
+            <TuitionAndHousingEstimates />
           </div>
           <div className="usa-width-two-thirds ">
             <div className="vads-l-row vads-u-flex-wrap--wrap">
@@ -133,17 +34,3 @@ export function SearchResults({
     </>
   );
 }
-const mapStateToProps = state => ({
-  eligibility: state.eligibility,
-  filters: state.filters,
-});
-
-const mapDispatchToProps = {
-  dispatchShowModal: showModal,
-  dispatchUpdateEligibilityAndFilters: updateEligibilityAndFilters,
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(SearchResults);
